requirements.txt
anyio @ file:///tmp/build/80754af9/anyio_1617783277988/work/dist
appdirs==1.4.4
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037499734/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
autopep8==1.6.0
awscli==1.22.98
Babel @ file:///tmp/build/80754af9/babel_1620871417480/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
botocore==1.24.43
branca==0.4.2
brotlipy==0.7.0
Cartopy==0.19.0.post1
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1613246951236/work
cftime @ file:///tmp/build/80754af9/cftime_1621823341290/work
chardet @ file:///tmp/build/80754af9/chardet_1607706775000/work
colorama==0.4.3
cryptography @ file:///tmp/build/80754af9/cryptography_1616767007030/work
cycler==0.10.0
Cython==0.29.23
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
docutils==0.15.2
entrypoints==0.3
folium==0.12.1.post1
fonttools==4.31.2
geojson==2.5.0
geojsoncontour==0.4.0
h11==0.13.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.9.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877307939/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1607452791405/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1622679242581/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1611333729159/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
jmespath==1.0.0
json5==0.9.5
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213314396/work
jupyter-packaging @ file:///tmp/build/80754af9/jupyter-packaging_1613502826984/work
jupyter-server @ file:///tmp/build/80754af9/jupyter_server_1616084066671/work
jupyterlab @ file:///tmp/build/80754af9/jupyterlab_1619133235951/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1617134334258/work
kaleido==0.2.1
kiwisolver==1.3.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621523467000/work
matplotlib==3.5.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
metar==1.9.0
MetPy==1.2.0
mistune @ file:///tmp/build/80754af9/mistune_1607364877025/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1618853971975/work
mkl-service==2.3.0
mplleaflet==0.0.5
nbclassic @ file:///tmp/build/80754af9/nbclassic_1616085367084/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1607370433589/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
netCDF4 @ file:///tmp/build/80754af9/netcdf4_1613684182594/work
notebook @ file:///tmp/build/80754af9/notebook_1621625621706/work
numexpr==2.8.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1620831195661/work
olefile==0.46
outcome==1.1.0
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.2.4
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120906940/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
Pint==0.18
plotly==5.6.0
pooch==1.6.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1618088486455/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
pygrib==2.1.4
pykdtree==1.3.4
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyproj==3.3.0
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1607365177477/work
pyshp==2.1.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
pyzmq==20.0.0
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
rsa==4.7.2
s3transfer==0.5.2
scipy @ file:///tmp/build/80754af9/scipy_1618855955483/work
selenium==4.1.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
Shapely==1.7.1
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1605205306277/work
sniffio @ file:///tmp/build/80754af9/sniffio_1614030464178/work
sortedcontainers==2.4.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
tables==3.7.0
tenacity==8.0.1
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
toml==0.10.2
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
trio==0.20.0
trio-websocket==0.9.2
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
wrapt==1.12.1
wrf-python==1.3.1
wsproto==1.1.0
xarray==0.18.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
